Fix seq_monadic hash narrowing warnings (#10541)
`seq_monadic` used 64-bit FNV constants directly in `size_t` hash
accumulators, which triggers constant-conversion warnings on 32-bit
`size_t` targets.

- **Hash accumulator width**
- Use `uint64_t` for the FNV-style accumulator in `group_sig_hash` and
product-state `key_hash`.
  - Keep the existing hash sequence unchanged.

- **Return boundary**
- Convert to `size_t` explicitly when returning from the hash functors.

```cpp
uint64_t h = 1469598103934665603ull;
// ...
return static_cast<size_t>(h);
```

+++ src/ast/rewriter/seq_monadic.h
@@ -195,12 +195,12 @@ private:
     typedef std::vector<std::pair<unsigned, unsigned>> group_sig;
     struct group_sig_hash {
         size_t operator()(group_sig const& s) const {
-            size_t h = 1469598103934665603ull;
+            uint64_t h = 1469598103934665603ull;
             for (auto const& p : s) {
                 h = (h ^ p.first) * 1099511628211ull;
                 h = (h ^ p.second) * 1099511628211ull;
             }
-            return h;
+            return static_cast<size_t>(h);
         }
     };
     group_sig m_sig_buf;                    // reused by group_nonempty (avoids allocating per lookup)
